+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-21 17:53:51
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-07-21 18:37:12
+// Commit: chore(state): update dependencies and cleanup unused imports in state
 
 class stateComponentState {
-  final String commitRef = '3007';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-07-21T17:53:51+05:00');
+  final String commitRef = '3011';
+  final String lastMessage = 'chore(state): update dependencies and cleanup unused imports in state';
+  final DateTime timestamp = DateTime.parse('2026-07-21T18:37:12+05:00');
 }